+++ src/components/home/MemeListContainer/type.ts
@@ -1,0 +1,1 @@
+export type MemeListType = "user" | "share" | "recent";
