+++ src/components/home/index.ts
@@ -1,3 +1,3 @@
-export * from "./MemeList";
+export * from "./MemeListContainer";
 export * from "./PopularTag";
 export * from "./UserSharedMeme";
